clean & split out base components (#8)
