+++ src/components/base/renderTextField/index.js
@@ -1,0 +1,19 @@
+import React, { PropTypes } from 'react';
+import TextField from 'material-ui/TextField';
+
+const renderTextField = ({ input, label, meta: { touched, error }, ...custom }) => (
+    <TextField  hintText={label}
+                floatingLabelText={label}
+                fullWidth={true}
+                errorText={touched && error}
+                {...input}
+                {...custom} />
+);
+
+renderTextField.propTypes = {
+    input: PropTypes.object,
+    label: PropTypes.string,
+    meta: PropTypes.object
+};
+
+export default renderTextField;
